refacto: fix vars and rename file
